[FIX] Fix some warnings caused by type mismatching in utringbuffer.

+++ src/utringbuffer.h
@@ -94,10 +94,10 @@ typedef struct {
  
 #define _utringbuffer_real_idx(a,j) ((a)->f ? ((j) + (a)->i) % (a)->n : (j))
 #define _utringbuffer_internalptr(a,j) ((void*)((a)->d + ((a)->icd.sz * (j))))
-#define utringbuffer_eltptr(a,j) ((0 <= (j) && (j) < utringbuffer_len(a)) ? _utringbuffer_internalptr(a,_utringbuffer_real_idx(a,j)) : NULL)
+#define utringbuffer_eltptr(a,j) ((0 <= (signed)(j) && (j) < utringbuffer_len(a)) ? _utringbuffer_internalptr(a,_utringbuffer_real_idx(a,j)) : NULL)
  
-#define _utringbuffer_fake_idx(a,j) ((a)->f ? ((j) + (a)->n - (a)->i) % (a)->n : (j))
-#define _utringbuffer_internalidx(a,e) (((char*)(e) >= (a)->d) ? (((char*)(e) - (a)->d)/(a)->icd.sz) : -1)
+#define _utringbuffer_fake_idx(a,j) ((a)->f ? ((j) + (a)->n - (a)->i) % (a)->n : (unsigned)(j))
+#define _utringbuffer_internalidx(a,e) (((char*)(e) >= (a)->d) ? (signed)(((char*)(e) - (a)->d)/(a)->icd.sz) : -1)
 #define utringbuffer_eltidx(a,e) _utringbuffer_fake_idx(a, _utringbuffer_internalidx(a,e))
  
 #define utringbuffer_front(a) utringbuffer_eltptr(a,0)
